+++ src/Resources/Contact/Contact.ts
@@ -1,4 +1,6 @@
 import AbstractResource from '../AbstractResource';
+import { HttpClientConfig, HttpClientResponse } from '../../Interfaces/HttpClient';
+import ContactForm from '../../Interfaces/contact/ContactForm';
 
 export default class Contact extends AbstractResource {
     initialise(): void {
@@ -6,7 +8,10 @@ export default class Contact extends AbstractResource {
         this.defaultAccess = 'guest';
     }
 
-    public submit: Function = this.createMethodFromPartialSpec({
+    public submit: (
+        formFields: ContactForm,
+        ...args: HttpClientConfig
+    ) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'POST',
         path: '/contact-us',
     });